calender project in c
This program prompts the user to enter a year and a month (as integers), and then calculates the number of days
in the month and the starting day of the week (using a modified version of the Zeller's congruence algorithm). 
It then prints a formatted calendar for the given month and year.

I will like to note though, that this program assumes the Gregorian calendar and does not handle leap seconds, leap years before 1582, or calendar reforms in different countries. Also, that  the formatting may vary depending on the console or terminal used to run the program.