Electricity Bill Calculator in c programming
So basically, this program first prompts the user to enter the total units consumed. 
Then it uses a series of if-else statements to calculate the total bill based on the following rates:

For the first 50 units, the rate is 0.50 per unit.
For the next 100 units (51 to 150), so  rate is 0.75 per unit.
For the next 100 units (151 to 250), the rate is 1.20 per unit.
For units above 250, the rate is 1.50 per unit.
In addition, a fixed charge of 50 is added to the total bill. Finally, the program prints the total bill to the console with two decimal places.